trading_bot/requirements.txt
ccxt
pandas
numpy
ta-lib
scikit-learn
tensorflow
flask
pyyaml
cryptography
requests
websockets